// src/api/models/getApiV10PostFilterBy.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * KOSMO Backend API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */

export type GetApiV10PostFilterBy = typeof GetApiV10PostFilterBy[keyof typeof GetApiV10PostFilterBy];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiV10PostFilterBy = {
  CLIENT: 'CLIENT',
  ADMIN: 'ADMIN',
} as const;