// src/api/models/getApiV10PostPosition.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * KOSMO Backend API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */

export type GetApiV10PostPosition = typeof GetApiV10PostPosition[keyof typeof GetApiV10PostPosition];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiV10PostPosition = {
  true: 'true',
  false: 'false',
} as const;