// src/api/models/getApiV10PostSortOrderPosition.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * KOSMO Backend API
 * Generated API documentation
 * OpenAPI spec version: 1.0.0
 */

export type GetApiV10PostSortOrderPosition = typeof GetApiV10PostSortOrderPosition[keyof typeof GetApiV10PostSortOrderPosition];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiV10PostSortOrderPosition = {
  ASC: 'ASC',
  DESC: 'DESC',
} as const;